Windows Tablet GPO's for Kiosk Mode
These GPO's lock the Windows tablet down (we implemented these with Microsoft Surface Pro 3's and 4's) into a Kiosk mode and automatically log the eScreening Service Account into Windows when powered on and automatically launches internet explorer to the default eScreening website. The GPO's will not allow the user to power the tablet down, modify the browser or anything else other than use the eScreening application. 

These GPO's should be given to your local IT department or Regional Service Line to modify/implement at the local site level. There is a GPO being applied that points to a batch file (eScreen.bat) located in the same folder this file is in that must be placed on the c:\ drive of the Windows tablet to function properly. The instructions for the tablet configuration (including directions on placing the batch file in the correct location) can be found in the documentation section of the Mental-Health-eScreening downloaded files (<site>-prod-release folder on