Correct the end date of a VistA appt string
When no appointments are returned, the code cannot get the visit string
for the appointment and so  it creates a new visit string based upon the
patient’s clinic location (where the assessment is taking place) and the
assessment date/time.  This new visit string will most likely be
different from the actual appointment visit string, and so when the TIU
note is created using the new visit string VistA will create an ad-hoc
visit for the note rather than associating the note with the patient’s
actual encounter.

The first problem with this code is with the use of Calendar.HOUR, which
is for 12-hour based times.  It seems that 24-hour time is intended
here, and so Calendar.HOUR_OF_DAY should be used.